refactor: define code to create SVG mapping file in a function
